+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * error pages changed;
 * `config.json` dropped;
 * ESLint errors fixed;
+* `npm run dev` task changed to run `lint`, `nodemon` and `gulp` concurrently;
 * `Dashboard` - link to `Facade` added;
 * `Facade/Gallery`:
     * photos navigation inside photo sets (prev/next);